pyproject.toml
[tool.poetry]
name = "core"
version = "0.0.1"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = {extras = ["asyncio", "mypy"], version = "^2.0.35"}
asyncpg = "^0.29.0"
pydantic = "^2.9.2"
alembic = "^1.13.3"
grpcio = "^1.66.2"
grpcio-reflection = "^1.66.2"
pydantic-settings = "^2.5.2"
grpcio-tools = "^1.66.2"
protobuf = "5.27.2"


[tool.poetry.group.dev.dependencies]
ruff = "^0.6.9"
mypy = "^1.11.2"
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
pytest-vcr = "^1.0.2"
grpcio-tools = "^1.66.2"
mypy-protobuf = "^3.6.0"
testcontainers = "^4.8.1"
pytest-env = "^1.1.5"
pytest-xdist = "^3.6.1"
pytest-anyio = "^0.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry-exec-plugin.commands]
test = "pytest -s -vv -n 1 --capture=tee-sys" #FIXME add tests
lint = """
    poetry run ruff check --fix \
    && poetry run ruff format \
    && poetry run mypy .
"""
# for local development
check = "ruff check --fix && ruff format ; mypy . && pytest -s -n 2"